+++ api/src/shared/database/database.py
@@ -21,6 +21,7 @@ from shared.database_gen.sqlacodegen_models import (
 from sqlalchemy.orm import sessionmaker
 import logging
 
+
 from shared.common.logging_utils import get_env_logging_level
 
 
@@ -116,78 +117,6 @@ def mapper_configure_listener(mapper, class_):
 event.listen(mapper, "mapper_configured", mapper_configure_listener)
 
 
-def create_refresh_materialized_view_task():
-    """
-    Asynchronously refresh a materialized view.
-    Ensures deduplication by generating a unique task name.
-
-    Returns:
-        dict: Response message and status code.
-    """
-    from google.protobuf import timestamp_pb2
-    from datetime import datetime, timedelta
-    from google.cloud import tasks_v2
-
-    try:
-        logging.info("Creating materialized view refresh task.")
-        now = datetime.now()
-
-        # BOUNCE WINDOW: next :00 or :30
-        minute = now.minute
-        if minute < 30:
-            bucket_time = now.replace(minute=30, second=0, microsecond=0)
-        else:
-            bucket_time = now.replace(minute=0, second=0, microsecond=0) + timedelta(hours=1)
-
-        timestamp_str = bucket_time.strftime("%Y-%m-%d-%H-%M")
-        task_name = f"refresh-materialized-view-{timestamp_str}"
-
-        # Convert to protobuf timestamp
-        proto_time = timestamp_pb2.Timestamp()
-        proto_time.FromDatetime(bucket_time)
-
-        # Cloud Tasks setup
-        client = tasks_v2.CloudTasksClient()
-        project = os.getenv("PROJECT_ID")
-        location = os.getenv("LOCATION")
-        queue = os.getenv("MATERIALIZED_VIEW_QUEUE")
-        url = (
-            f"https://{os.getenv('GCP_REGION')}-"
-            f"{os.getenv('PROJECT_ID')}.cloudfunctions.net/"
-            f"tasks-executor-{os.getenv('ENVIRONMENT_NAME')}"
-        )
-
-        task_name = client.task_path(project, location, queue, task_name)
-
-        # Convert to protobuf timestamp
-        proto_time = timestamp_pb2.Timestamp()
-        proto_time.FromDatetime(bucket_time)
-
-        # Enqueue the task
-        try:
-            create_http_task_with_name(
-                client=client,
-                body=b"",
-                url=url,
-                project_id=project,
-                gcp_region=location,
-                queue_name=queue,
-                task_name=task_name,
-                task_time=proto_time,
-                http_method=tasks_v2.HttpMethod.GET,
-            )
-            logging.info(f"Scheduled refresh materialized view task for {timestamp_str}")
-            return {"message": f"Refresh task for {timestamp_str} scheduled."}, 200
-        except Exception as e:
-            if "ALREADY_EXISTS" in str(e):
-                logging.info(f"Task already exists for {timestamp_str}, skipping.")
-
-    except Exception as error:
-        error_msg = f"Error enqueuing task: {error}"
-        logging.error(error_msg)
-        return {"error": error_msg}, 500
-
-
 def refresh_materialized_view(session: "Session", view_name: str) -> bool:
     """
     Refresh Materialized view by name.
@@ -199,47 +128,6 @@ def refresh_materialized_view(session: "Session", view_name: str) -> bool:
     except Exception as error:
         logging.error("Error raised while refreshing view: %s", error)
         return False
-
-
-def create_http_task_with_name(
-    client,  # type: tasks_v2.CloudTasksClient
-    body: bytes,
-    url: str,
-    project_id: str,
-    gcp_region: str,
-    queue_name: str,
-    task_name: str,
-    task_time: str,
-    http_method,
-):
-    from google.cloud import tasks_v2
-
-    """Creates a GCP Cloud Task."""
-
-    token = (tasks_v2.OidcToken(service_account_email=os.getenv("SERVICE_ACCOUNT_EMAIL")),)
-
-    task = tasks_v2.Task(
-        http_request=tasks_v2.HttpRequest(
-            url=url,
-            http_method=http_method,
-            oidc_token=token,
-            body=body,
-            headers={"Content-Type": "application/json"},
-        )
-    )
-    task = {
-        "name": task_name,
-        "http_request": {
-            "http_method": http_method,
-            "url": url,
-            "headers": {
-                "Content-Type": "application/json",
-                "Authorization": f"Bearer {token}",
-            },
-        },
-        "schedule_time": task_time,
-    }
-    client.create_task(parent=client.queue_path(project_id, gcp_region, queue_name), task=task)
 
 
 def with_db_session(func=None, db_url: str | None = None):
