+++ functions-python/process_validation_report/requirements.txt
@@ -16,6 +16,7 @@ geoalchemy2==0.14.7
 
 # Google specific packages for this function
 cloudevents~=1.10.1
+google-cloud-tasks
 
 # Configuration
 python-dotenv==1.0.0
